+++ datalad/interface/__init__.py
@@ -81,8 +81,7 @@ _group_misc = (
         ('datalad.distributed.export_to_figshare', 'ExportToFigshare'),
         ('datalad.interface.test', 'Test'),
         ('datalad.interface.clean', 'Clean'),
-        ('datalad.interface.add_archive_content', 'AddArchiveContent',
-         'add-archive-content'),
+        ('datalad.local.add_archive_content', 'AddArchiveContent'),
         ('datalad.interface.download_url', 'DownloadURL', 'download-url'),
         ('datalad.interface.shell_completion', 'ShellCompletion', 'shell-completion'),
         ('datalad.core.local.run', 'Run', 'run'),
